breaking shoebox-scraper client out of shoebox general (kitchen sink) client. both client and controllers are very large and its hard to understand who is using what and deal with timeouts in these scopes. next would be to break the huge controller and extract few calls like with the ShoeboxDataPipeController. the new service client has higher timeouts to reduce noise and be more tolerant for these calls
