+++ kifi-backend/modules/common/app/com/keepit/shoebox/ShoeboxServiceClient.scala
@@ -33,12 +33,9 @@ import com.keepit.common.usersegment.UserSegmentKey
 import com.keepit.common.cache.TransactionalCaching.Implicits.directCacheAccess
 import com.keepit.heimdal.SanitizedKifiHit
 import com.keepit.model.serialize.{ UriIdAndSeqBatch, UriIdAndSeq }
-import org.msgpack.ScalaMessagePack
-import com.keepit.common.time.RichDateTime
 
 trait ShoeboxServiceClient extends ServiceClient {
   final val serviceType = ServiceType.SHOEBOX
-  private val ? = null
 
   def getUserOpt(id: ExternalId[User]): Future[Option[User]]
   def getSocialUserInfoByNetworkAndSocialId(id: SocialId, networkType: SocialNetworkType): Future[Option[SocialUserInfo]]
@@ -87,23 +84,6 @@ trait ShoeboxServiceClient extends ServiceClient {
   def getNormalizedUriUpdates(lowSeq: SequenceNumber[ChangedURI], highSeq: SequenceNumber[ChangedURI]): Future[Seq[(Id[NormalizedURI], NormalizedURI)]]
   def kifiHit(clicker: Id[User], hit: SanitizedKifiHit): Future[Unit]
   def getHelpRankInfos(uriIds: Seq[Id[NormalizedURI]]): Future[Seq[HelpRankInfo]]
-  def assignScrapeTasks(zkId: Long, max: Int): Future[Seq[ScrapeRequest]]
-  def isUnscrapableP(url: String, destinationUrl: Option[String]): Future[Boolean]
-  def isUnscrapable(url: String, destinationUrl: Option[String]): Future[Boolean]
-  def getLatestKeep(url: String): Future[Option[Keep]]
-  def getBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]): Future[Seq[Keep]]
-  def saveBookmark(bookmark: Keep): Future[Keep]
-  def saveScrapeInfo(info: ScrapeInfo): Future[Unit]
-  def saveNormalizedURI(uri: NormalizedURI): Future[NormalizedURI]
-  def updateNormalizedURIState(uriId: Id[NormalizedURI], state: State[NormalizedURI]): Future[Unit]
-  def savePageInfo(pageInfo: PageInfo): Future[Unit]
-  def getImageInfo(id: Id[ImageInfo]): Future[ImageInfo]
-  def saveImageInfo(imgInfo: ImageInfo): Future[ImageInfo]
-  def updateNormalizedURI(uriId: => Id[NormalizedURI], createdAt: => DateTime = ?, updatedAt: => DateTime = ?, externalId: => ExternalId[NormalizedURI] = ?, title: => Option[String] = ?, url: => String = ?, urlHash: => UrlHash = UrlHash(?), state: => State[NormalizedURI] = ?, seq: => SequenceNumber[NormalizedURI] = SequenceNumber(-1), screenshotUpdatedAt: => Option[DateTime] = ?, restriction: => Option[Restriction] = ?, normalization: => Option[Normalization] = ?, redirect: => Option[Id[NormalizedURI]] = ?, redirectTime: => Option[DateTime] = ?): Future[Unit]
-  def recordPermanentRedirect(uri: NormalizedURI, redirect: HttpRedirect): Future[NormalizedURI]
-  def recordScrapedNormalization(uriId: Id[NormalizedURI], uriSignature: Signature, candidateUrl: String, candidateNormalization: Normalization, alternateUrls: Set[String]): Future[Unit]
-  def getProxy(url: String): Future[Option[HttpProxy]]
-  def getProxyP(url: String): Future[Option[HttpProxy]]
   def getFriendRequestsBySender(senderId: Id[User]): Future[Seq[FriendRequest]]
   def getUserValue(userId: Id[User], key: UserValueName): Future[Option[String]]
   def setUserValue(userId: Id[User], key: UserValueName, value: String): Unit
@@ -222,24 +202,6 @@ class ShoeboxServiceClientImpl @Inject() (
       call(Shoebox.internal.getBookmarkByUriAndUser(uriId, userId)).map { r =>
         Json.fromJson[Option[Keep]](r.json).get
       }
-    }
-  }
-
-  def getBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]): Future[Seq[Keep]] = {
-    call(Shoebox.internal.getBookmarksByUriWithoutTitle(uriId), callTimeouts = longTimeout).map { r =>
-      r.json.as[JsArray].value.map(js => Json.fromJson[Keep](js).get)
-    }
-  }
-
-  def getLatestKeep(url: String): Future[Option[Keep]] = {
-    call(Shoebox.internal.getLatestKeep(), callTimeouts = longTimeout, body = JsString(url)).map { r =>
-      Json.fromJson[Option[Keep]](r.json).get
-    }
-  }
-
-  def saveBookmark(bookmark: Keep): Future[Keep] = {
-    call(Shoebox.internal.saveBookmark(), Json.toJson(bookmark), callTimeouts = longTimeout).map { r =>
-      Json.fromJson[Keep](r.json).get
     }
   }
 
@@ -567,134 +529,6 @@ class ShoeboxServiceClientImpl @Inject() (
     val payload = Json.toJson(uriIds)
     call(Shoebox.internal.getHelpRankInfo, payload) map { r =>
       r.json.as[Seq[HelpRankInfo]]
-    }
-  }
-
-  def assignScrapeTasks(zkId: Long, max: Int): Future[Seq[ScrapeRequest]] = {
-    call(Shoebox.internal.assignScrapeTasks(zkId, max), callTimeouts = longTimeout, routingStrategy = leaderPriority).map { r =>
-      r.json.as[Seq[ScrapeRequest]]
-    }
-  }
-
-  def saveScrapeInfo(info: ScrapeInfo): Future[Unit] = {
-    call(Shoebox.internal.saveScrapeInfo(), Json.toJson(info), callTimeouts = longTimeout).map { r => Unit }
-  }
-
-  def savePageInfo(pageInfo: PageInfo): Future[Unit] = {
-    call(Shoebox.internal.savePageInfo(), Json.toJson(pageInfo), callTimeouts = longTimeout, routingStrategy = leaderPriority).map { r => Unit }
-  }
-
-  def getImageInfo(id: Id[ImageInfo]): Future[ImageInfo] = {
-    call(Shoebox.internal.getImageInfo(id)).map { r =>
-      r.json.as[ImageInfo]
-    }
-  }
-
-  def saveImageInfo(imgInfo: ImageInfo): Future[ImageInfo] = {
-    call(Shoebox.internal.saveImageInfo(), Json.toJson(imgInfo), callTimeouts = longTimeout, routingStrategy = leaderPriority).map { r =>
-      r.json.as[ImageInfo]
-    }
-  }
-
-  @deprecated("Dangerous call. Use updateNormalizedURI instead.", "2014-01-30")
-  def saveNormalizedURI(uri: NormalizedURI): Future[NormalizedURI] = {
-    call(Shoebox.internal.saveNormalizedURI(), Json.toJson(uri), callTimeouts = longTimeout).map { r =>
-      r.json.as[NormalizedURI]
-    }
-  }
-
-  def updateNormalizedURIState(uriId: Id[NormalizedURI], state: State[NormalizedURI]): Future[Unit] = {
-    val json = Json.obj("state" -> state)
-    call(Shoebox.internal.updateNormalizedURI(uriId), json, callTimeouts = longTimeout).imap(_ => {})
-  }
-
-  def updateNormalizedURI(uriId: => Id[NormalizedURI],
-    createdAt: => DateTime,
-    updatedAt: => DateTime,
-    externalId: => ExternalId[NormalizedURI],
-    title: => Option[String],
-    url: => String,
-    urlHash: => UrlHash,
-    state: => State[NormalizedURI],
-    seq: => SequenceNumber[NormalizedURI],
-    screenshotUpdatedAt: => Option[DateTime],
-    restriction: => Option[Restriction],
-    normalization: => Option[Normalization],
-    redirect: => Option[Id[NormalizedURI]],
-    redirectTime: => Option[DateTime]): Future[Unit] = {
-    import com.keepit.common.strings.OptionWrappedJsObject
-    val safeUrlHash = Option(urlHash).map(p => Option(p.hash)).flatten
-    val safeSeq = Option(seq).map(v => if (v.value == -1L) None else Some(v)).flatten
-
-    val safeJsonParams: Seq[(String, JsValueWrapper)] = Seq(
-      "createdAt" -> Option(createdAt),
-      "updatedAt" -> Option(updatedAt),
-      "externalId" -> Option(externalId),
-      "title" -> Option(title),
-      "url" -> Option(url),
-      "urlHash" -> safeUrlHash,
-      "state" -> Option(state),
-      "seq" -> safeSeq,
-      "screenshotUpdatedAt" -> Option(screenshotUpdatedAt),
-      "restriction" -> Option(restriction),
-      "normalization" -> Option(normalization),
-      "redirect" -> Option(redirect),
-      "redirectTime" -> Option(redirectTime)
-    )
-    val payload = Json.obj(safeJsonParams: _*)
-    val stripped = payload.stripJsNulls()
-    call(Shoebox.internal.updateNormalizedURI(uriId), stripped, callTimeouts = longTimeout, routingStrategy = leaderPriority).imap(_ => {})
-  }
-
-  def recordPermanentRedirect(uri: NormalizedURI, redirect: HttpRedirect): Future[NormalizedURI] = {
-    call(Shoebox.internal.recordPermanentRedirect(), JsArray(Seq(Json.toJson[NormalizedURI](uri), Json.toJson[HttpRedirect](redirect))), callTimeouts = longTimeout).map { r =>
-      r.json.as[NormalizedURI]
-    }
-  }
-
-  def recordScrapedNormalization(uriId: Id[NormalizedURI], uriSignature: Signature, candidateUrl: String, candidateNormalization: Normalization, alternateUrls: Set[String]): Future[Unit] = {
-    val payload = Json.obj(
-      "id" -> uriId.id,
-      "signature" -> uriSignature.toBase64(),
-      "url" -> candidateUrl,
-      "normalization" -> candidateNormalization,
-      "alternateUrls" -> alternateUrls
-    )
-    call(Shoebox.internal.recordScrapedNormalization(), payload, callTimeouts = longTimeout).imap(_ => {})
-  }
-
-  def getProxy(url: String): Future[Option[HttpProxy]] = {
-    call(Shoebox.internal.getProxy(url)).map { r =>
-      if (r.json == null) None else r.json.asOpt[HttpProxy]
-    }
-  }
-
-  def getProxyP(url: String): Future[Option[HttpProxy]] = {
-    call(Shoebox.internal.getProxyP, Json.toJson(url), callTimeouts = longTimeout).map { r =>
-      if (r.json == null) None else r.json.asOpt[HttpProxy]
-    }
-  }
-
-  def isUnscrapable(url: String, destinationUrl: Option[String]): Future[Boolean] = {
-    call(Shoebox.internal.isUnscrapable(url.take(MaxUrlLength), destinationUrl.map(_.take(MaxUrlLength)))).map { r =>
-      r.json.as[Boolean]
-    }
-  }
-
-  def isUnscrapableP(url: String, destinationUrl: Option[String]): Future[Boolean] = {
-    val destUrl = if (destinationUrl.isDefined && url == destinationUrl.get) {
-      log.debug(s"[isUnscrapableP] url==destUrl ${url}; ignored") // todo: fix calling code
-      None
-    } else destinationUrl map { dUrl =>
-      log.debug(s"[isUnscrapableP] url($url) != destUrl($dUrl)")
-      dUrl
-    }
-    val payload = JsArray(destUrl match {
-      case Some(dUrl) => Seq(Json.toJson(url.take(MaxUrlLength)), Json.toJson(dUrl.take(MaxUrlLength)))
-      case None => Seq(Json.toJson(url.take(MaxUrlLength)))
-    })
-    call(Shoebox.internal.isUnscrapableP, payload, callTimeouts = longTimeout).map { r =>
-      r.json.as[Boolean]
     }
   }
