+++ kifi-backend/modules/scraper/app/com/keepit/dev/ScraperDevModule.scala
@@ -1,13 +1,13 @@
 package com.keepit.dev
 
-import com.keepit.common.cache.ScraperCacheModule
+import com.keepit.common.cache.{ HashMapMemoryCacheModule, ScraperCacheModule }
+import com.keepit.common.concurrent.DevForkJoinContextMonitorModule
+import com.keepit.common.store.ScraperDevStoreModule
+import com.keepit.inject.CommonDevModule
+import com.keepit.scraper.embedly.DevEmbedlyModule
 import com.keepit.scraper.fetcher.DevHttpFetcherModule
 import com.keepit.scraper.{ DevScraperProcessorModule, ScraperServiceModule }
-import com.keepit.common.cache.HashMapMemoryCacheModule
-import com.keepit.inject.CommonDevModule
-import com.keepit.common.store.ScraperDevStoreModule
-import com.keepit.common.concurrent.DevForkJoinContextMonitorModule
-import com.keepit.scraper.embedly.DevEmbedlyModule
+import com.keepit.shoebox.{ ProdShoeboxScraperClientModule, ProdShoeboxServiceClientModule }
 
 case class ScraperDevModule() extends ScraperServiceModule(
   cacheModule = ScraperCacheModule(HashMapMemoryCacheModule()),
@@ -16,5 +16,7 @@ case class ScraperDevModule() extends ScraperServiceModule(
   scrapeProcessorModule = DevScraperProcessorModule(),
   embedlyModule = DevEmbedlyModule()
 ) with CommonDevModule {
+  val shoeboxServiceClientModule = ProdShoeboxServiceClientModule()
+  val shoeboxScraperClientModule = ProdShoeboxScraperClientModule()
   val fetcherModule = DevHttpFetcherModule()
 }
