+++ kifi-backend/modules/scraper/app/com/keepit/scraper/ShoeboxDbCallbacks.scala
@@ -1,7 +1,7 @@
 package com.keepit.scraper
 
 import com.google.inject.{ Inject, Singleton }
-import com.keepit.shoebox.ShoeboxServiceClient
+import com.keepit.shoebox.{ ShoeboxScraperClient, ShoeboxServiceClient }
 import scala.concurrent.{ Future, Await, Awaitable }
 import com.keepit.model._
 import com.keepit.common.db.{ State, Id }
@@ -11,7 +11,12 @@ import java.util.concurrent.locks.ReentrantLock
 import com.keepit.common.logging.Logging
 
 @Singleton
-class ShoeboxDbCallbackHelper @Inject() (config: ScraperConfig, shoeboxServiceClient: ShoeboxServiceClient) extends SyncShoeboxDbCallbacks with ShoeboxDbCallbacks with Logging {
+class ShoeboxDbCallbackHelper @Inject() (
+  config: ScraperConfig,
+  shoeboxServiceClient: ShoeboxServiceClient,
+  shoeboxScraperClient: ShoeboxScraperClient)
+    extends SyncShoeboxDbCallbacks with ShoeboxDbCallbacks with Logging {
+
   implicit val serviceCallTimeout = config.serviceCallTimeout
   implicit val fjCtx = ExecutionContext.fj
 
@@ -65,25 +70,25 @@ class ShoeboxDbCallbackHelper @Inject() (config: ScraperConfig, shoeboxServiceCl
     await(recordScrapedNormalization(uriId, uriSignature, candidateUrl, candidateNormalization, alternateUrls))
   }
 
-  def assignTasks(zkId: Long, max: Int): Future[Seq[ScrapeRequest]] = shoeboxServiceClient.assignScrapeTasks(zkId, max)
+  def assignTasks(zkId: Long, max: Int): Future[Seq[ScrapeRequest]] = shoeboxScraperClient.assignScrapeTasks(zkId, max)
   def getNormalizedUri(uri: NormalizedURI): Future[Option[NormalizedURI]] = {
     uri.id match {
       case Some(id) => shoeboxServiceClient.getNormalizedURI(id).map(Some(_))
       case None => shoeboxServiceClient.getNormalizedURIByURL(uri.url)
     }
   }
-  def saveNormalizedUri(uri: NormalizedURI): Future[NormalizedURI] = shoeboxServiceClient.saveNormalizedURI(uri)
-  def updateNormalizedURIState(uriId: Id[NormalizedURI], state: State[NormalizedURI]): Future[Unit] = shoeboxServiceClient.updateNormalizedURIState(uriId, state)
-  def saveScrapeInfo(info: ScrapeInfo): Future[Unit] = shoeboxServiceClient.saveScrapeInfo(if (info.state == ScrapeInfoStates.INACTIVE) info else info.withState(ScrapeInfoStates.ACTIVE))
-  def savePageInfo(info: PageInfo): Future[Unit] = shoeboxServiceClient.savePageInfo(info)
-  def saveImageInfo(info: ImageInfo): Future[ImageInfo] = shoeboxServiceClient.saveImageInfo(info)
-  def getBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]): Future[Seq[Keep]] = shoeboxServiceClient.getBookmarksByUriWithoutTitle(uriId)
-  def getLatestKeep(url: String): Future[Option[Keep]] = shoeboxServiceClient.getLatestKeep(url)
-  def saveBookmark(bookmark: Keep): Future[Keep] = shoeboxServiceClient.saveBookmark(bookmark)
-  def recordPermanentRedirect(uri: NormalizedURI, redirect: HttpRedirect): Future[NormalizedURI] = shoeboxServiceClient.recordPermanentRedirect(uri, redirect)
-  def isUnscrapableP(url: String, destinationUrl: Option[String]) = shoeboxServiceClient.isUnscrapableP(url, destinationUrl)
+  def saveNormalizedUri(uri: NormalizedURI): Future[NormalizedURI] = shoeboxScraperClient.saveNormalizedURI(uri)
+  def updateNormalizedURIState(uriId: Id[NormalizedURI], state: State[NormalizedURI]): Future[Unit] = shoeboxScraperClient.updateNormalizedURIState(uriId, state)
+  def saveScrapeInfo(info: ScrapeInfo): Future[Unit] = shoeboxScraperClient.saveScrapeInfo(if (info.state == ScrapeInfoStates.INACTIVE) info else info.withState(ScrapeInfoStates.ACTIVE))
+  def savePageInfo(info: PageInfo): Future[Unit] = shoeboxScraperClient.savePageInfo(info)
+  def saveImageInfo(info: ImageInfo): Future[ImageInfo] = shoeboxScraperClient.saveImageInfo(info)
+  def getBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]): Future[Seq[Keep]] = shoeboxScraperClient.getBookmarksByUriWithoutTitle(uriId)
+  def getLatestKeep(url: String): Future[Option[Keep]] = shoeboxScraperClient.getLatestKeep(url)
+  def saveBookmark(bookmark: Keep): Future[Keep] = shoeboxScraperClient.saveBookmark(bookmark)
+  def recordPermanentRedirect(uri: NormalizedURI, redirect: HttpRedirect): Future[NormalizedURI] = shoeboxScraperClient.recordPermanentRedirect(uri, redirect)
+  def isUnscrapableP(url: String, destinationUrl: Option[String]) = shoeboxScraperClient.isUnscrapableP(url, destinationUrl)
   def recordScrapedNormalization(uriId: Id[NormalizedURI], uriSignature: Signature, candidateUrl: String, candidateNormalization: Normalization, alternateUrls: Set[String]): Future[Unit] = {
-    shoeboxServiceClient.recordScrapedNormalization(uriId, uriSignature, candidateUrl, candidateNormalization, alternateUrls)
+    shoeboxScraperClient.recordScrapedNormalization(uriId, uriSignature, candidateUrl, candidateNormalization, alternateUrls)
   }
   override def updateURIRestriction(uriId: Id[NormalizedURI], r: Option[Restriction]): Unit = await({
     shoeboxServiceClient.updateURIRestriction(uriId, r)
